tag_generator/requirements.txt
beautifulsoup4==4.7.1
bitly-api==0.3
certifi==2019.3.9
chardet==3.0.4
cssselect==1.0.3
feedfinder2==0.0.4
feedparser==5.2.1
idna==2.8
jieba3k==0.35.1
lxml==4.3.4
newspaper3k==0.2.8
nltk==3.2.4
oauthlib==3.0.1
ParallelDots==3.2.13
Pillow==6.0.0
preprocessing==0.1.13
PyMySQL==0.9.3
PySocks==1.7.0
python-dateutil==2.8.0
PyYAML==5.1.1
requests==2.22.0
requests-file==1.4.3
requests-oauthlib==1.2.0
six==1.12.0
soupsieve==1.9.1
sphinx-rtd-theme==0.2.4
textblob==0.15.3
tinysegmenter==0.3
tldextract==2.2.1
tweepy==3.7.0
urllib3==1.25.3
